backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
faiss-cpu
sentence-transformers
rank_bm25